+++ src/types.rs
@@ -126,17 +126,17 @@ unsafe impl<'a, T: IntoValue + FromValue<'a>> FromValue<'a> for Array<'a, T> {
 
 impl<'a> Array<'a, f64> {
     /// Set value to double array
-    pub fn set_double(&mut self, i: usize, f: f64) -> Result<(), Error> {
+    pub fn set_f64(&mut self, i: usize, f: f64) -> Result<(), Error> {
         if i >= self.len() {
             return Err(CamlError::ArrayBoundError.into());
         }
 
-        if !self.is_double_array() {
+        if !self.is_f64() {
             return Err(Error::NotDoubleArray);
         }
 
         unsafe {
-            self.set_double_unchecked(i, f);
+            self.set_f64_unchecked(i, f);
         };
 
         Ok(())
@@ -147,21 +147,20 @@ impl<'a> Array<'a, f64> {
     /// # Safety
     /// This function performs no bounds checking
     #[inline]
-    pub unsafe fn set_double_unchecked(&mut self, i: usize, f: f64) {
-        let ptr = ((self.0).raw().0 as *mut f64).add(i);
-        *ptr = f;
+    pub unsafe fn set_f64_unchecked(&mut self, i: usize, f: f64) {
+        self.0.store_f64_field(i, f)
     }
 
     /// Get a value from a double array
-    pub fn get_double(&self, i: usize) -> Result<f64, Error> {
+    pub fn get_f64(&self, i: usize) -> Result<f64, Error> {
         if i >= self.len() {
             return Err(CamlError::ArrayBoundError.into());
         }
-        if !self.is_double_array() {
+        if !self.is_f64() {
             return Err(Error::NotDoubleArray);
         }
 
-        Ok(unsafe { self.get_double_unchecked(i) })
+        Ok(unsafe { self.get_f64_unchecked(i) })
     }
 
     /// Get a value from a double array without checking if the array is actually a double array
@@ -170,8 +169,8 @@ impl<'a> Array<'a, f64> {
     ///
     /// This function does not perform bounds checking
     #[inline]
-    pub unsafe fn get_double_unchecked(&self, i: usize) -> f64 {
-        *(self.0.raw().0 as *mut f64).add(i)
+    pub unsafe fn get_f64_unchecked(&self, i: usize) -> f64 {
+        self.0.f64_field(i)
     }
 }
 
@@ -182,9 +181,9 @@ impl<'a, T: IntoValue + FromValue<'a>> Array<'a, T> {
         Array(x, PhantomData)
     }
 
-    /// Check if Array contains only doubles, if so `get_double` and `set_double` should be used
+    /// Check if Array contains only doubles, if so `get_f64` and `set_f64` should be used
     /// to access values
-    pub fn is_double_array(&self) -> bool {
+    pub fn is_f64(&self) -> bool {
         unsafe { sys::caml_is_double_array(self.0.raw().0) == 1 }
     }